tests/wpt/web-platform-tests/css/css-text/parsing/word-boundary-detection-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing word-boundary-detection with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#word-boundary-detection">
<meta name="assert" content="word-boundary-detection supports the full grammar 'normal | manual | auto(<lang>) '.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("word-boundary-detection", "normal");
test_valid_value("word-boundary-detection", "manual");
/* Not sure how to test for auto(<lang>) generically, as which one will parse depends on languages actually supported */
</script>
</body>
</html>
